--- compiler/inc/ANSI.h
@@ -1,72 +1,0 @@
-// Regular Colors
-#define RC_Black "\033[0;30m"
-#define RC_Red "\033[0;31m"
-#define RC_Green "\033[0;32m"
-#define RC_Yellow "\033[0;33m"
-#define RC_Blue "\033[0;34m"
-#define RC_Purple "\033[0;35m"
-#define RC_Cyan "\033[0;36m"
-#define RC_White "\033[0;37m"
-
-// Bold
-#define B_Black "\033[1;30m"
-#define B_Red "\033[1;31m"
-#define B_Green "\033[1;32m"
-#define B_Yellow "\033[1;33m"
-#define B_Blue "\033[1;34m"
-#define B_Purple "\033[1;35m"
-#define B_Cyan "\033[1;36m"
-#define B_White "\033[1;37m"
-
-// Underline
-#define UL_Black "\033[4;30m"
-#define UL_Red "\033[4;31m"
-#define UL_Green "\033[4;32m"
-#define UL_Yellow "\033[4;33m"
-#define UL_Blue "\033[4;34m"
-#define UL_Purple "\033[4;35m"
-#define UL_Cyan "\033[4;36m"
-#define UL_White "\033[4;37m"
-
-// Background
-#define BG_Black "\033[40m"
-#define BG_Red "\033[41m"
-#define BG_Green "\033[42m"
-#define BG_Yellow "\033[43m"
-#define BG_Blue "\033[44m"
-#define BG_Purple "\033[45m"
-#define BG_Cyan "\033[46m"
-#define BG_White "\033[47m"
-
-// High Intensty
-#define HI_Black "\033[0;90m"
-#define HI_Red "\033[0;91m"
-#define HI_Green "\033[0;92m"
-#define HI_Yellow "\033[0;93m"
-#define HI_Blue "\033[0;94m"
-#define HI_Purple "\033[0;95m"
-#define HI_Cyan "\033[0;96m"
-#define HI_White "\033[0;97m"
-
-// Bold High Intensty
-#define BHI_Black "\033[1;90m"
-#define BHI_Red "\033[1;91m"
-#define BHI_Green "\033[1;92m"
-#define BHI_Yellow "\033[1;93m"
-#define BHI_Blue "\033[1;94m"
-#define BHI_Purple "\033[1;95m"
-#define BHI_Cyan "\033[1;96m"
-#define BHI_White "\033[1;97m"
-
-// High Intensty backgrounds
-#define HIBG_Black "\033[0;100m"
-#define HIBG_Red "\033[0;101m"
-#define HIBG_Green "\033[0;102m"
-#define HIBG_Yellow "\033[0;103m"
-#define HIBG_Blue "\033[0;104m"
-#define HIBG_Purple "\033[0;105m"
-#define HIBG_Cyan "\033[0;106m"
-#define HIBG_White "\033[0;107m"
-
-// Reset
-#define Reset "\033[0m"
