+++ compiler/inc/parser.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "scanner.h"
+
+typedef struct
+{
+
+} AST;
+
+
+auto parse(std::vector<token> tokens) -> AST;
