--- compiler/main.c
@@ -1,7 +1,0 @@
-#include "inc/lexer.h"
-
-int main(int argc, char **argv)
-{
-    parse(argv[1]); 
-    return 0;
-}
