+++ compiler/src/error_handler.cpp
@@ -1,0 +1,12 @@
+#include "../inc/error_handler.h"
+
+void report(bool condition, std::string error_message, bool critical)
+{
+    if (condition)
+    {
+        std::cerr << error_message << std::endl;
+        if (critical)
+            exit(EXIT_FAILURE);
+    }
+}
+
